added constant auton speed and consolidated code in Robot class
